requirements.txt
openai
langchain
langchain_openai
langchain_experimental
langchain_community
langchainhub
pinecone-client
python-dotenv
tiktoken
docx2txt
pypdf
requests
numpy
pandas
duckduckgo-search
wikipedia